requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.4b2
bleach==6.0.0
braceexpand==0.1.7
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.2.0
cmake==3.26.3
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.5
crc32c==2.2.post0
cycler==0.11.0
cytoolz==0.11.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dm-pix==0.2.1
dm-tree==0.1.8
docstring-parser==0.14.1
easydict==1.9
etils==1.2.0
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
filelock==3.12.0
fire==0.4.0
flatbuffers==2.0
fonttools==4.37.3
freetype-py==2.3.0
frozendict==2.3.2
frozenlist==1.3.0
fsspec==2023.4.0
future==0.18.2
fvcore==0.1.5.post20220414
gast==0.4.0
gin-config==0.5.0
google-api-core==2.7.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
h5py==3.6.0
httplib2==0.20.4
hydra-core==1.1.2
idna==3.3
imageio==2.22.0
imgaug==0.4.0
importlib-metadata==4.12.0
importlib-resources==5.2.3
iopath==0.1.9
ipycanvas==0.12.0
ipydatawidgets==4.3.3
ipyevents==2.0.1
ipygany==0.5.0
ipykernel==6.22.0
ipython==8.12.1
ipython-genutils==0.2.0
ipyvtklink==0.2.2
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.17.3
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.5.0
kubernetes==18.20.0
libclang==13.0.0
lit==16.0.2
locket==0.2.0
lpips==0.1.4
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdurl==0.1.2
meshcat==0.3.2
mistune==2.0.5
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.6
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
omegaconf==2.1.2
open3d==0.13.0
opencv-python==4.5.5.64
opencv-python-headless==4.7.0.72
openmesh==1.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
plyfile==0.7.4
portalocker==2.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
pyglet==1.5.24
Pygments==2.15.0
pymeshlab==2022.2.post2
pyngrok==5.1.0
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.9
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.19.3
python-dateutil==2.8.2
pythreejs==2.4.2
pytorch-lightning==1.2.9
pytz==2023.3
pyvista==0.34.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==11.1.0
rsa==4.8
scikit-image==0.17.2
scikit-learn==1.0.2
scipy==1.8.1
scooby==0.5.12
seaborn==0.12.2
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.6.2
strip-hints==0.1.10
sympy==1.11.1
tabulate==0.8.9
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-probability==0.19.0
tensorstore==0.1.35
termcolor==1.1.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
tfp-nightly==0.17.0.dev20220722
tfrecord==1.14.1
threadpoolctl==3.1.0
threevis==0.1.0.post25
tifffile==2022.8.12
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.10.2
toolz==0.12.0
torch==2.0.0
torchmetrics==0.2.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.64.0
traitlets
traittypes==0.2.1
transforms3d==0.4.1
trimesh==3.12.0
triton==2.0.0
typer==0.4.1
typing-extensions==4.5.0
tzdata==2023.3
u-msgpack-python==2.7.1
uritemplate==3.0.1
urllib3==1.26.9
vtk==9.1.0
wcwidth==0.2.6
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
widgetsnbextension==4.0.7
wldhx.yadisk-direct==0.0.6
wrapt==1.14.0
wslink==1.6.4
yacs==0.1.8
yarl==1.7.2
zipp==3.8.1